Revised. One issue remained (adding geom to links). Working on it
