+++ ra2ce/graph/networks.py
@@ -21,20 +21,16 @@
 
 import logging
 import os
-from pathlib import Path
-from typing import Any, List, Tuple, Union
+from typing import Any, List, Tuple
 
 import geopandas as gpd
 import networkx as nx
-import osmnx
 import pandas as pd
 import pyproj
-from networkx import Graph
-from shapely.geometry import MultiLineString, Polygon, MultiPolygon
-from shapely.geometry.base import BaseGeometry
+from shapely.geometry import MultiLineString
 
 import ra2ce.graph.networks_utils as nut
-from ra2ce.graph.osm_network_wrapper import OsmNetworkWrapper
+from ra2ce.osm_network_wrapper.osm_network_wrapper import OsmNetworkWrapper
 from ra2ce.graph.segmentation import Segmentation
 from ra2ce.io.readers import GraphPickleReader
 from ra2ce.io.writers import JsonExporter
